Remove DomainsStore from DomainManagement.Nameservers
After switching over to remove the call to `sites-list` in the
nameserver configuration component we started receiving an error when
navigating back from the DNS settings to the nameserver config panel.

This was caused by a store listener in the `<StoreConnection />`
component which was causing `setState()` _inside_ of the `render()`
cycle.

To fix this we had to stop listening to the domain store and start using
the domain selectors with the `<QuerySiteDomains />` component.

There is now no longer any error.